Move `applyEnterprisePlugin` to AbstractCrossVersionPerformanceTest
